Add fail if no sequences test
